Create CollapsedStoryLabels component
