+++ app/assets/javascripts/components/story/CollapsedStory/CollapsedStoryInfo.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import CollapsedStoryTitle from './CollapsedStoryTitle'
-import { labelSplit } from '../../../rules/story';
+import CollapsedStoryLabels from './CollapsedStoryLabels'
 
 const CollapsedStoryInfo = ({ title, labels, ownedByName, ownedByInitials }) => (
   <div className="Story__info">
-    <StoryLabels labels={labels} />
+    <CollapsedStoryLabels labels={labels} />
     <CollapsedStoryTitle title={title} ownedByInitials={ownedByInitials} ownedByName={ownedByName} />
   </div>
 );
@@ -17,28 +17,6 @@ CollapsedStoryInfo.propTypes = {
 
 CollapsedStoryInfo.defaultProps = {
   labels: '',
-};
-
-const StoryLabel = ({ label }) => (
-  <a href="#" className="Story__label" title={label}>{label}</a>
-);
-
-StoryLabel.propTypes = {
-  label: PropTypes.string.isRequired,
-};
-
-const StoryLabels = ({ labels }) => {
-  if (!labels) {
-    return null
-  }
-
-  return (
-    <span className='Story__labels'>
-      {labelSplit(labels).map(label => (
-        <StoryLabel key={label} label={label} />
-      ))}
-    </span>
-  );
 };
 
 export default CollapsedStoryInfo;
