+++ app/assets/javascripts/components/story/CollapsedStory/CollapsedStoryLabels.js
@@ -1,0 +1,27 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import { labelSplit } from '../../../rules/story';
+
+const StoryLabel = ({ label }) => (
+  <a href="#" className="Story__label" title={label}>{label}</a>
+);
+
+StoryLabel.propTypes = {
+  label: PropTypes.string.isRequired,
+};
+
+const CollapsedStoryLabels = ({ labels }) => {
+  if (!labels) {
+    return null
+  }
+
+  return (
+    <span className='Story__labels'>
+      {labelSplit(labels).map(label => (
+        <StoryLabel key={label} label={label} />
+      ))}
+    </span>
+  );
+};
+
+export default CollapsedStoryLabels;
